Builder constructor for FeedForward
